Parse "coverage" runner argument.
This makes AndroidJUnitRunner work with gradle JaCoCo support.

I'm not sure how to test this. I think the best way is to have a unit
test that reflectively checks that a result of fromBundle has every
field set to something other than default. This would force new fields
to be added to fromBundle. What do you think?

Change-Id: Iccac6744bd43642ca7d2a2f687215241ac54031d