Ignore suite()s and don't ignore init errors when using method filters.
JUnit3 TestSuite's do not support filters. There is an android specific work
around in place, but unfortunately it cannot help when tests create their own
suites via a suite() method. This leads to inconsistent behaviour with
InstrumentationTestRunner when a class has a suite() method, but runner was
instructed to only run a single method. This change will instruct the builder
to ignore suite() methods when a method filter has been provided.

Also add a special case to method filter to never filter out 'initializationError'
descriptions, since those are generated on error cases.

Change-Id: I402d43b0e5d2731512fadae7e9003eb46a5402e5