Add special case multidex installation for ICS and lower.
Typically multidex is installed by inserting call at Application#attachBaseContext
However in  ICS and presumably below, instrumentation#onCreate is called before
attachBaseContext, leading to potential issues if instrumentation#onCreate code path
tries to load any classes that are not in primary dex.

This commit adds a reflective call to Multidex.install in onCreate when running on a
api <= ICS. It should have no effect if multidex is not on the classpath.

Change-Id: I319e207bfbbaede9e3b01cc981878d330e49b82a