Added a view action to perform given action repeatedly on a view until view matches given view matcher.
Added repeatedlyUntil() method in ViewActions which takes action to be performed repeatedly until the given matcher matches the view on which this action is performed. Integration test is also added for this view action.
Worked on the review feedback
Re-setting the Espresso failure handler in a couple of tests to DefaultFailureHandler to avoid its side effect on other tests

Change-Id: Id3b8521453464e1d78091fea4793788042483dd3